back_end/requirements.txt
Flask==1.1.2
Flask_Cors==4.0.0
protobuf==5.26.1
PyPDF2==3.0.1
python-dotenv==1.0.1
python_docx==1.1.0
